Refactor parsed body mapping to its own middleware
