+++ chocs/middleware/__init__.py
@@ -1,3 +1,4 @@
 from .application_middleware import ApplicationMiddleware
 from .middleware import Middleware, MiddlewareCursor, MiddlewareFunction, MiddlewareHandler, MiddlewarePipeline
 from .open_api_middleware import OpenApiMiddleware
+from .parsed_body_middleware import ParsedBodyMiddleware
