+++ chocs/middleware/open_api_middleware.py
@@ -1,15 +1,14 @@
-import inspect
-from typing import Callable, Dict, Any, List, Optional, Union
+from typing import Callable, Dict, List, Optional, Union
 
+from chocs.http_message import JsonHttpMessage
 from chocs.http_method import HttpMethod
 from chocs.http_request import HttpRequest
 from chocs.http_response import HttpResponse
+from chocs.json_schema.errors import InvalidInputValidationError
 from chocs.json_schema.json_schema import JsonReference, OpenApiSchema
 from chocs.json_schema.schema_validator import build_validator_from_schema
-from chocs.json_schema.errors import InvalidInputValidationError
 from chocs.middleware import Middleware, MiddlewareHandler
 from chocs.routing import Route
-from chocs.http_message import JsonHttpMessage
 
 
 class OpenApiMiddleware(Middleware):
@@ -56,39 +55,7 @@ class OpenApiMiddleware(Middleware):
         else:
             raise InvalidInputValidationError(message="Request body could not be validated.")
 
-        valid_body = body_validator(parsed_body)
-
-        if "parsed_body" not in route.attributes:
-            return
-
-        if not inspect.isclass(route.attributes["parsed_body"]):
-            return
-
-        strict = route.attributes["strict"] if "strict" in route.attributes else True
-        constructor = route.attributes["parsed_body"]
-        request._parsed_body = None
-
-        if not strict:
-
-            def _get_non_strict_parsed_body() -> Any:
-
-                instance = constructor.__new__(constructor)
-                for prop_name, prop_value in valid_body.items():
-                    setattr(instance, prop_name, prop_value)
-
-                if hasattr(instance, "__post_init__"):
-                    instance.__post_init__()
-
-                return instance
-
-            request._parsed_body_getter = _get_non_strict_parsed_body
-
-            return
-
-        def _get_strict_parsed_body() -> Any:
-            return constructor(**valid_body)
-
-        request._parsed_body_getter = _get_strict_parsed_body
+        request._parsed_body = body_validator(parsed_body)
 
     def _get_query_validator(self, route: str, method: HttpMethod) -> Optional[Callable]:
         if route not in self.query_validators:
