+++ tests/middleware/test_openapi_middleware.py
@@ -84,74 +84,6 @@ def test_turn_off_query_validation() -> None:
     assert app(HttpRequest(HttpMethod.GET, "/pets"))
 
 
-def test_hydrate_parsed_body_with_strict_mode() -> None:
-    app = _mockup_app()
-
-    @dataclass
-    class Pet:
-        name: str
-        tag: str
-        id: str
-
-    @app.post("/pets", parsed_body=Pet)
-    def create_pet(request: HttpRequest) -> HttpResponse:
-        pet = request.parsed_body  # type: Pet
-        assert isinstance(pet, Pet)
-        return HttpResponse(pet.name)
-
-    invalid_body = json.dumps({
-        "name": "Bobek",
-        "tag": "test",
-        "id": 1,
-        "unknown_property": "unknown_value"
-    })
-
-    with pytest.raises(TypeError):
-        app(HttpRequest(HttpMethod.POST, "/pets", body=invalid_body, headers={"content-type": "application/json"}))
-
-    valid_body = json.dumps({
-        "name": "Bobek",
-        "tag": "test",
-        "id": 1,
-    })
-    response = app(HttpRequest(HttpMethod.POST, "/pets", body=valid_body, headers={"content-type": "application/json"}))
-    assert str(response) == "Bobek"
-
-
-def test_hydrate_parsed_body_without_strict_mode() -> None:
-    app = _mockup_app()
-
-    @dataclass
-    class Pet:
-        name: str
-        tag: str
-        id: str
-
-    @app.post("/pets", parsed_body=Pet, strict=False)
-    def create_pet(request: HttpRequest) -> HttpResponse:
-        pet = request.parsed_body  # type: Pet
-        assert isinstance(pet, Pet)
-        return HttpResponse(pet.name)
-
-    invalid_body = json.dumps({
-        "name": "Bobek",
-        "tag": "test",
-        "id": 1,
-        "unknown_property": "unknown_value"
-    })
-
-    respone = app(HttpRequest(HttpMethod.POST, "/pets", body=invalid_body, headers={"content-type": "application/json"}))
-    assert str(respone) == "Bobek"
-
-    valid_body = json.dumps({
-        "name": "Bobek",
-        "tag": "test",
-        "id": 1,
-    })
-    response = app(HttpRequest(HttpMethod.POST, "/pets", body=valid_body, headers={"content-type": "application/json"}))
-    assert str(response) == "Bobek"
-
-
 def test_body_validation_handles_arrays() -> None:
     openapi_file = path.realpath(path.dirname(__file__) + "/../fixtures/openapi_request_body_array.yml")
     app = Application(OpenApiMiddleware(openapi_file, validate_body=True, validate_query=True))
